fix(nits):  renamed state folder to store, to better signify it's purpose
